Use json storage lib for prac

+++ bot/storage/json.js
@@ -1,23 +1,30 @@
 const fs = require('fs')
+const path = require('path')
 const moment = require('moment')
 const Storage = require('./storage')
 
 class Json extends Storage {
 
-    load(filename, defaultJson = []) {
+    load(filebase, defaultJson = []) {
         try {
-            return require(this.path(filename + '.json'))
+            return require(this.fileBase(this.path(filebase)) + '.json')
         } catch (error) {
             return defaultJson
         }
     }
 
-    save(filename, data) {
+    save(filebase, data) {
+        if (!filebase) {
+            return false
+        }
         try {
-            const tmpFile = this.path(filename + (new Date().getTime()) + '.json')
-            fs.copyFile(this.path(filename + '.json'), tmpFile, (cpErr) => {
-                fs.writeFile(this.path(filename + '.json'), JSON.stringify(data), 'utf8', (wErr) => {
-                    if (!wErr) {
+            filebase = this.fileBase(this.path(filebase))
+            const filePath = filebase + '.json'
+            this.folderCheck(filePath) // Make sure folder exists
+            const tmpFile = filebase + (new Date().getTime()) + '.json'
+            fs.copyFile(filePath, tmpFile, (cpErr) => {
+                fs.writeFile(filePath, JSON.stringify(data), 'utf8', (wErr) => {
+                    if (!wErr && !cpErr) {
                         fs.unlink(tmpFile, (dErr) => { })
                     }
                 })
@@ -26,6 +33,10 @@ class Json extends Storage {
         } catch (error) {
             return false
         }
+    }
+
+    fileBase(filePath) {
+        return path.join(this.folderName(filePath), path.basename(filePath, '.json'))
     }
 
 }
